DELIA-54869 :  License request retry logic Reason for change: add retry logic and handle jsonrpc failure case
Test Procedure: Will be updated in the ticket

Risks: Low

(cherry picked from commit ad3b27a967b6fc079ba823dd8e911dde6b2ee3a0)

+++ drm/AampSecManager.h
@@ -74,7 +74,7 @@ public:
 						const char* mediaUsage, const char* accessToken, size_t accessTokenLen,
 						int64_t* sessionId,
 						char** licenseResponse, size_t* licenseResponseLength,
-						int64_t* statusCode, int64_t* reasonCode);
+						int32_t* statusCode, int32_t* reasonCode);
 
 
 	/**
